+++ src/main/java/com/shortkki/api/recipeImport/port/RecipeParserPort.java
@@ -1,0 +1,7 @@
+package com.shortkki.api.recipeImport.port;
+
+import com.shortkki.api.recipeImport.dto.RecipeParseResult;
+
+public interface RecipeParserPort {
+    RecipeParseResult parseRecipeFromUrl(String url);
+}
