+++ src/main/java/com/shortkki/api/recipeImport/service/RecipeImportAsyncService.java
@@ -1,39 +1,21 @@
 package com.shortkki.api.recipeImport.service;
 
-import com.shortkki.api.ingredient.entity.Ingredient;
-import com.shortkki.api.ingredient.repository.IngredientRepository;
 import com.shortkki.api.member.entity.Member;
 import com.shortkki.api.member.repository.MemberRepository;
-import com.shortkki.api.recipe.constant.CuisineType;
-import com.shortkki.api.recipe.constant.Difficulty;
-import com.shortkki.api.recipe.constant.MealType;
 import com.shortkki.api.recipe.entity.Recipe;
-import com.shortkki.api.recipe.entity.RecipeIngredient;
-import com.shortkki.api.recipe.entity.RecipeStep;
-import com.shortkki.api.recipe.repository.RecipeIngredientRepository;
-import com.shortkki.api.recipe.repository.RecipeRepository;
-import com.shortkki.api.recipe.repository.RecipeStepRepository;
-import com.shortkki.api.recipeBook.entity.RecipeBook;
 import com.shortkki.api.recipeBook.service.RecipeBookQueryService;
 import com.shortkki.api.recipeBook.service.RecipeBookService;
 import com.shortkki.api.recipeImport.dto.RecipeParseResult;
-import com.shortkki.api.recipeImport.dto.RecipeParseResult.IngredientParseResult;
-import com.shortkki.api.recipeImport.dto.RecipeParseResult.StepParseResult;
-import com.shortkki.api.recipeImport.service.parser.AiRecipeParserService;
+import com.shortkki.api.recipeImport.port.RecipeParserPort;
 import com.shortkki.api.source.domain.SourceContent;
 import com.shortkki.api.source.domain.ImportStatus;
 import com.shortkki.api.source.domain.SourceImportHistory;
 import com.shortkki.api.source.repository.SourceContentRepository;
 import com.shortkki.api.source.repository.SourceImportHistoryRepository;
-import com.shortkki.api.recipe.entity.vo.RecipeBasicInfo;
-import com.shortkki.api.recipe.entity.vo.RecipeCategoryInfo;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
 
 @Slf4j
 @Service
@@ -42,15 +24,11 @@ public class RecipeImportAsyncService {
 
     private final SourceContentRepository sourceContentRepository;
     private final MemberRepository memberRepository;
-    private final RecipeRepository recipeRepository;
     private final RecipeBookService recipeBookService;
     private final RecipeBookQueryService recipeBookQueryService;
     private final SourceImportHistoryRepository sourceImportHistoryRepository;
-    private final AiRecipeParserService aiRecipeParserService;
-    private final IngredientRepository ingredientRepository;
-    private final RecipeIngredientRepository recipeIngredientRepository;
-    private final RecipeStepRepository recipeStepRepository;
-    private final PlatformTransactionManager transactionManager;
+    private final RecipeParserPort recipeParserPort;
+    private final RecipeImportTransactionalService transactionalService;
 
     @Async
     public void processImport(
@@ -82,7 +60,7 @@ public class RecipeImportAsyncService {
             history.updateStatus(ImportStatus.PARSING);
             sourceImportHistoryRepository.save(history);
             log.info("AI 레시피 파싱 시작: {}", sourceUrl);
-            parseResult = aiRecipeParserService.parseRecipeFromUrl(sourceUrl);
+            parseResult = recipeParserPort.parseRecipeFromUrl(sourceUrl);
             aiRawResponse = parseResult.rawResponse();
             log.info("AI 파싱 완료 - 제목: {}, 재료: {}개, 순서: {}개",
                     parseResult.title(),
@@ -96,17 +74,7 @@ public class RecipeImportAsyncService {
 
         Recipe saved;
         try {
-            saved = new TransactionTemplate(transactionManager).execute(status -> {
-                try {
-                    Recipe recipe = saveRecipe(member, sourceContent, parseResult);
-                    saveIngredients(recipe, parseResult.ingredients());
-                    saveSteps(recipe, parseResult.steps());
-                    return recipe;
-                } catch (Exception e) {
-                    status.setRollbackOnly();
-                    throw e;
-                }
-            });
+            saved = transactionalService.saveRecipeWithRelations(member, sourceContent, parseResult);
         } catch (Exception e) {
             log.error("Recipe save failed", e);
             history.fail("Save failed: " + e.getMessage());
@@ -130,112 +98,9 @@ public class RecipeImportAsyncService {
         addToDefaultRecipeBook(memberId, saved.getId());
     }
 
-    private Recipe saveRecipe(Member member, SourceContent sourceContent, RecipeParseResult parseResult) {
-        RecipeBasicInfo basicInfo = new RecipeBasicInfo(
-                parseResult.title() != null ? parseResult.title() : sourceContent.getTitle(),
-                parseResult.description(),
-                parseResult.servingSize() != null ? parseResult.servingSize() : 1,
-                parseResult.cookingTime() != null ? parseResult.cookingTime() : 0);
-
-        CuisineType cuisineType = parseCuisineType(parseResult.cuisineType());
-        MealType mealType = parseMealType(parseResult.mealType());
-        Difficulty difficulty = parseDifficulty(parseResult.difficulty());
-
-        RecipeCategoryInfo categoryInfo = new RecipeCategoryInfo(
-                cuisineType != null ? cuisineType : CuisineType.ETC,
-                mealType != null ? mealType : MealType.ETC,
-                difficulty != null ? difficulty : Difficulty.ETC);
-
-        Recipe recipe = Recipe.createImported(
-                member,
-                basicInfo,
-                categoryInfo,
-                sourceContent);
-        return recipeRepository.save(recipe);
-    }
-
-    private void saveIngredients(Recipe recipe, List<IngredientParseResult> ingredients) {
-        for (IngredientParseResult ing : ingredients) {
-            String unit = parseUnit(ing.amount());
-            Ingredient ingredient = ingredientRepository.findByName(ing.name())
-                    .orElseGet(() -> ingredientRepository.save(
-                            Ingredient.create(ing.name(), unit)));
-            Integer amount = parseAmount(ing.amount());
-            RecipeIngredient recipeIngredient = RecipeIngredient.create(
-                    ingredient, recipe, amount);
-            recipeIngredientRepository.save(recipeIngredient);
-        }
-    }
-
-    private Integer parseAmount(String amountStr) {
-        if (amountStr == null || amountStr.isBlank()) {
-            return null;
-        }
-        try {
-            String numberOnly = amountStr.replaceAll("[^0-9]", "");
-            return numberOnly.isEmpty() ? null : Integer.parseInt(numberOnly);
-        } catch (NumberFormatException e) {
-            return null;
-        }
-    }
-
-    private String parseUnit(String amountStr) {
-        if (amountStr == null || amountStr.isBlank()) {
-            return "";
-        }
-        String unit = amountStr.replaceAll("[0-9/\\.\\-]+", "").trim();
-        return unit.isBlank() ? "" : unit;
-    }
-
-    private void saveSteps(Recipe recipe, List<StepParseResult> steps) {
-        for (StepParseResult step : steps) {
-            RecipeStep recipeStep = RecipeStep.create(
-                    recipe, step.stepNumber(), step.description());
-            recipeStepRepository.save(recipeStep);
-        }
-    }
-
     private void addToDefaultRecipeBook(Long memberId, Long recipeId) {
-        recipeBookQueryService.findAllByMemberId(memberId).stream()
-                .filter(RecipeBook::getIsDefault)
-                .findFirst()
-                .ifPresent(defaultBook -> recipeBookService.addRecipeInternal(
+        recipeBookQueryService.findDefaultByMemberId(memberId)
+                .ifPresent(defaultBook -> recipeBookService.addRecipeIfNotExists(
                         memberId, defaultBook.getId(), recipeId));
-    }
-
-    private CuisineType parseCuisineType(String value) {
-        if (value == null) {
-            return null;
-        }
-        try {
-            return CuisineType.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown cuisineType: {}", value);
-            return null;
-        }
-    }
-
-    private MealType parseMealType(String value) {
-        if (value == null) {
-            return null;
-        }
-        try {
-            return MealType.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown mealType: {}", value);
-            return null;
-        }
-    }
-
-    private Difficulty parseDifficulty(String value) {
-        if (value == null) {
-            return null;
-        }
-        try {
-            return Difficulty.valueOf(value);
-        } catch (IllegalArgumentException e) {
-            log.warn("Unknown difficulty: {}", value);
-            return null;
-        }
     }
 }
